+++ lib/vm/src/lib.rs
@@ -39,7 +39,9 @@ pub mod libcalls;
 pub use crate::export::*;
 pub use crate::global::*;
 pub use crate::imports::Imports;
-pub use crate::instance::{ImportInitializerFuncPtr, InstanceAllocator, InstanceHandle};
+pub use crate::instance::{
+    ImportFunctionEnv, ImportInitializerFuncPtr, InstanceAllocator, InstanceHandle,
+};
 pub use crate::memory::{LinearMemory, Memory, MemoryError, MemoryStyle};
 pub use crate::mmap::Mmap;
 pub use crate::module::{ExportsIterator, ImportsIterator, ModuleInfo};